// src/shared/api/models/getApiPreviousSalesItemsCustomerId200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiPreviousSalesItemsCustomerId200ItemsItem } from './getApiPreviousSalesItemsCustomerId200ItemsItem';

export type GetApiPreviousSalesItemsCustomerId200 = {
    currentPage: number;
    limit: number;
    items: GetApiPreviousSalesItemsCustomerId200ItemsItem[];
};